Add an alias property named "authors"
This object class now introduces a property named "authors" that is an
alias for the "author_affiliations" property that Dimensions uses.
The reason for creating a field name that does not exist in Dimensions
is that I found "author_affiliations" too confusing (because it's not
just the affiliations -- it's the actual author list too) and,
simultaneously, the lack of an "authors" field really unintuitive for
a publication.  Still, creating two ways of accessing the same data is
not ideal and may introduce code maintenance issues in the future.
Time will tell.