Improve dealing with rate limits
This now implements a back-off and retry strategy if the rate limit is
hit despite our (separate) attempts to track API usage and pause when
we know we've counted enough calls.  It also fixes the per-minute rate
limit number to what I empirically determined is closer to what the
Dimensions server actually seems to impose (which is 23 calls per
minute and not 30 per minute).